LASB-1287: Moved the court data adapter client package from laastatus into it's own courtdataadapter package.
