+++ maat-court-data-api/src/main/java/gov/uk/courtdata/courtdataadapter/client/CourtDataAdapterOAuth2ClientConfig.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.laastatus.client;
+package gov.uk.courtdata.courtdataadapter.client;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
