+++ maat-court-data-api/src/main/java/gov/uk/courtdata/laastatus/service/LaaStatusPostCDAService.java
@@ -3,7 +3,7 @@ package gov.uk.courtdata.laastatus.service;
 import com.amazonaws.xray.spring.aop.XRayEnabled;
 import gov.uk.courtdata.dto.CourtDataDTO;
 import gov.uk.courtdata.laastatus.builder.RepOrderUpdateMessageBuilder;
-import gov.uk.courtdata.laastatus.client.CourtDataAdapterClient;
+import gov.uk.courtdata.courtdataadapter.client.CourtDataAdapterClient;
 import gov.uk.courtdata.model.laastatus.LaaStatusUpdate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
